refactor: separate async tasks from handlers
