+++ internal/queue/client_integration_test.go
@@ -1,4 +1,4 @@
-package asynq_test
+package queue_test
 
 import (
 	"context"
@@ -10,15 +10,16 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"github.com/opcotech/elemo/internal/config"
+	"github.com/opcotech/elemo/internal/queue"
 	"github.com/opcotech/elemo/internal/testutil"
-	elemoAsynq "github.com/opcotech/elemo/internal/transport/asynq"
+	elemoAsynq "github.com/opcotech/elemo/internal/transport/async"
 )
 
 type AsynqClientIntegrationTestSuite struct {
 	testutil.ContainerIntegrationTestSuite
 	testutil.RedisContainerIntegrationTestSuite
 
-	client *elemoAsynq.Client
+	client *queue.Client
 	worker *elemoAsynq.Worker
 }
 
@@ -31,8 +32,8 @@ func (s *AsynqClientIntegrationTestSuite) SetupSuite() {
 
 	var err error
 
-	s.client, err = elemoAsynq.NewClient(
-		elemoAsynq.WithClientConfig(&config.WorkerConfig{
+	s.client, err = queue.NewClient(
+		queue.WithClientConfig(&config.WorkerConfig{
 			Concurrency: 10,
 			Broker:      s.RedisConf.RedisConfig,
 		}),
@@ -48,7 +49,7 @@ func (s *AsynqClientIntegrationTestSuite) SetupSuite() {
 			Concurrency: 10,
 			Broker:      s.RedisConf.RedisConfig,
 		}),
-		elemoAsynq.WithWorkerTaskHandler(elemoAsynq.TaskTypeSystemHealthCheck, systemHealthCheckTaskHandler),
+		elemoAsynq.WithWorkerTaskHandler(queue.TaskTypeSystemHealthCheck, systemHealthCheckTaskHandler),
 	)
 	s.Require().NoError(err)
 
@@ -73,7 +74,7 @@ func (s *AsynqClientIntegrationTestSuite) TestEnqueue() {
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	task, err := elemoAsynq.NewSystemHealthCheckTask()
+	task, err := queue.NewSystemHealthCheckTask()
 	s.Require().NoError(err)
 
 	info, err := s.client.Enqueue(ctx, task)
@@ -87,7 +88,7 @@ func (s *AsynqClientIntegrationTestSuite) TestGetTaskInfo() {
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	task, err := elemoAsynq.NewSystemHealthCheckTask()
+	task, err := queue.NewSystemHealthCheckTask()
 	s.Require().NoError(err)
 
 	info, err := s.client.Enqueue(ctx, task)
