+++ internal/queue/client_test.go
@@ -1,12 +1,8 @@
-package asynq
+package queue
 
 import (
-	"context"
-	"reflect"
 	"testing"
 
-	"github.com/hibiken/asynq"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/opcotech/elemo/internal/config"
@@ -15,14 +11,14 @@ import (
 	"github.com/opcotech/elemo/internal/testutil/mock"
 )
 
-func TestWithWorkerConfig(t *testing.T) {
+func TestWithClientConfig(t *testing.T) {
 	type args struct {
 		config *config.WorkerConfig
 	}
 	tests := []struct {
 		name    string
 		args    args
-		want    *Worker
+		want    *Client
 		wantErr error
 	}{
 		{
@@ -30,7 +26,7 @@ func TestWithWorkerConfig(t *testing.T) {
 			args: args{
 				config: new(config.WorkerConfig),
 			},
-			want: &Worker{
+			want: &Client{
 				conf: new(config.WorkerConfig),
 			},
 		},
@@ -46,71 +42,17 @@ func TestWithWorkerConfig(t *testing.T) {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			worker := new(Worker)
-			err := WithWorkerConfig(tt.args.config)(worker)
+			client := new(Client)
+			err := WithClientConfig(tt.args.config)(client)
 			require.ErrorIs(t, err, tt.wantErr)
 			if tt.wantErr == nil {
-				require.Equal(t, tt.want, worker)
+				require.Equal(t, tt.want, client)
 			}
 		})
 	}
 }
 
-func TestWithWorkerTaskHandler(t *testing.T) {
-	handler := asynq.HandlerFunc(func(_ context.Context, _ *asynq.Task) error {
-		return nil
-	})
-
-	type args struct {
-		taskType TaskType
-		handler  asynq.Handler
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    *Worker
-		wantErr error
-	}{
-		{
-			name: "create new option with handler",
-			args: args{
-				taskType: TaskTypeSystemHealthCheck,
-				handler:  handler,
-			},
-			want: &Worker{
-				handlers: map[TaskType]asynq.Handler{
-					TaskTypeSystemHealthCheck: handler,
-				},
-			},
-		},
-		{
-			name: "create new option with nil handler",
-			args: args{
-				taskType: TaskTypeSystemHealthCheck,
-				handler:  nil,
-			},
-			wantErr: ErrNoTaskHandler,
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			worker := new(Worker)
-			worker.handlers = make(map[TaskType]asynq.Handler)
-
-			err := WithWorkerTaskHandler(tt.args.taskType, tt.args.handler)(worker)
-			require.ErrorIs(t, err, tt.wantErr)
-			if tt.wantErr == nil {
-				for k, v := range tt.want.handlers {
-					assert.Equal(t, reflect.ValueOf(v).Pointer(), reflect.ValueOf(worker.handlers[k]).Pointer())
-				}
-			}
-		})
-	}
-}
-
-func TestWithWorkerLogger(t *testing.T) {
+func TestWithClientLogger(t *testing.T) {
 	type args struct {
 		logger log.Logger
 	}
@@ -139,15 +81,15 @@ func TestWithWorkerLogger(t *testing.T) {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			worker := new(Worker)
-			err := WithWorkerLogger(tt.args.logger)(worker)
+			client := new(Client)
+			err := WithClientLogger(tt.args.logger)(client)
 			require.ErrorIs(t, err, tt.wantErr)
-			require.Equal(t, tt.want, worker.logger)
+			require.Equal(t, tt.want, client.logger)
 		})
 	}
 }
 
-func TestWithWorkerTracer(t *testing.T) {
+func TestWithClientTracer(t *testing.T) {
 	type args struct {
 		tracer tracing.Tracer
 	}
@@ -176,10 +118,10 @@ func TestWithWorkerTracer(t *testing.T) {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			worker := new(Worker)
-			err := WithWorkerTracer(tt.args.tracer)(worker)
+			client := new(Client)
+			err := WithClientTracer(tt.args.tracer)(client)
 			require.ErrorIs(t, err, tt.wantErr)
-			require.Equal(t, tt.want, worker.tracer)
+			require.Equal(t, tt.want, client.tracer)
 		})
 	}
 }
