+++ internal/transport/async/handler.go
@@ -1,43 +1,16 @@
-package asynq
+package async
 
 import (
-	"time"
-
 	"github.com/opcotech/elemo/internal/pkg/log"
 	"github.com/opcotech/elemo/internal/pkg/tracing"
 	"github.com/opcotech/elemo/internal/service"
 )
 
-const (
-	DefaultTaskTimeout   = 5 * time.Second // The default task timeout.
-	DefaultTaskRetention = 5 * time.Minute // The default task retention.
-)
-
-const (
-	TaskTypeSystemHealthCheck   TaskType = iota + 1 // Health check task type.
-	TaskTypeSystemLicenseExpiry                     // License expiry task type.
-)
-
-var (
-	taskTypeValues = map[TaskType]string{
-		TaskTypeSystemHealthCheck:   "system:health_check",
-		TaskTypeSystemLicenseExpiry: "system:license_expiry",
-	}
-)
-
-// TaskType is the type for system tasks.
-type TaskType uint8
-
-// String returns the string representation of the system task type.
-func (t TaskType) String() string {
-	return taskTypeValues[t]
-}
-
-// TaskOption is a function that can be used to configure a task handler.
-type TaskOption func(*baseTaskHandler) error
+// TaskHandlerOption is a function that can be used to configure a task handler.
+type TaskHandlerOption func(*baseTaskHandler) error
 
 // WithTaskEmailService sets the email service for the worker.
-func WithTaskEmailService(emailService service.EmailService) TaskOption {
+func WithTaskEmailService(emailService service.EmailService) TaskHandlerOption {
 	return func(t *baseTaskHandler) error {
 		if emailService == nil {
 			return ErrNoEmailService
@@ -49,7 +22,7 @@ func WithTaskEmailService(emailService service.EmailService) TaskOption {
 }
 
 // WithTaskLogger sets the logger for the task handler.
-func WithTaskLogger(logger log.Logger) TaskOption {
+func WithTaskLogger(logger log.Logger) TaskHandlerOption {
 	return func(t *baseTaskHandler) error {
 		if logger == nil {
 			return log.ErrNoLogger
@@ -62,7 +35,7 @@ func WithTaskLogger(logger log.Logger) TaskOption {
 }
 
 // WithTaskTracer sets the tracer for the task handler.
-func WithTaskTracer(tracer tracing.Tracer) TaskOption {
+func WithTaskTracer(tracer tracing.Tracer) TaskHandlerOption {
 	return func(t *baseTaskHandler) error {
 		if tracer == nil {
 			return tracing.ErrNoTracer
@@ -83,7 +56,7 @@ type baseTaskHandler struct {
 }
 
 // newBaseTaskHandler creates a new base task handler.
-func newBaseTaskHandler(opts ...TaskOption) (*baseTaskHandler, error) {
+func newBaseTaskHandler(opts ...TaskHandlerOption) (*baseTaskHandler, error) {
 	t := &baseTaskHandler{
 		logger: log.DefaultLogger(),
 		tracer: tracing.NoopTracer(),
