+++ internal/transport/async/middleware_test.go
@@ -1,4 +1,4 @@
-package asynq
+package async
 
 import (
 	"context"
